Incorperated Request, Certification and connection to API
Still needs some modification to make sure that we get values from from the API. Most of this code is reliant on the code made within the Bitbucket reference from 23_example.

Need to think some more about how the values will be returned to main, so that everything will run as smoothly as possible.